Add an InfoCard to various listy pages
This is very handy in itself, by being inline documentation, but
it also handily moves our items down the screen a bit, so people
don't have to reach quite so far. We are supposed to be able to
pull the lists down as well, but that is not currently working(?),
so until that works, this already helps with that.