Clear the queue when disabling casual mode (to avoid misfires)
This does cause potential removal of e.g. alarms which have are due
in the next thirty seconds (for example), which isn't great, but
as we have had tails falling off tables, this would be the lesser
of those two evils. Be good to fix this a little more pleasantly
later on.