Ensure we disable commands after they've been active for too long
Hackery hacky time - if we end up running for longer than we're supposed to,
assume that we missed something, and set ourselves available again.
Also, we may end up in a situation where we have, in fact, been activated
and deactivated correctly. In that case, let's not deactivate things we need
to keep active. This could be done above, but keeping the code together feels
simpler for future maintenance.