Handle the occasionally odd notifications from the tail
We will occasionally get END COMMANDBEGIN OTHERCOMM back from
the tail rather than two separate END and BEGIN notifications,
with the notification we get truncated to the btle message
value length limit, so we do not, in fact, have sufficient
data to reasonably handle what's going on. So, we reintroduce
the hack, and make sure to mention why this time.