Disable casual mode when enabling gestures and vice versa
This is an unfortunate situation, and it would be nice to be able to
revert this at some point, but for now it's the fastest way to ensure
that we don't end up burning out servos by just continuously hammering
them with commands. Kind of a problem, so... here we are.