Rework the bt permission requests somewhat, combine them into one
This is a step towards hopefully making it less awkward to handle the
bluetooth scan and connect settings we've got on Android. It removes the
flow logic from qml (where it really should have never been) and puts it
into the permissions manager itself (which is now less generic, but more
useful).