Remove gear from device list (by deleting it) when it's shut down
This is handled explicitly for the SHUTDOWN command, but also listens to
the low energy controller disconnect signal, and it then deletes itself
if it didn't happen due to firmware updating stuff (at which point we'll
want to keep the device around, as we're about to reconnect to it).

We also increase the reconnector timeout to 100 ms instead of next event
loop run, as it's kind of awkward to just spam those messages, and it's
usually a case of "heyyyy waiiiit iiiii'm geeeetttttiiiinggg tooo iiit",
so it makes sense to just hang back a bit there.