Add in-progress notification permission
This is not needed yet (we don't need it technically for when we only
use the notification as the foreground service keepalive thing), but
it'd be nice to have for sure. However, we will definitely want this at
some point, because as ToeiRei points out, actually using this for
something other than just saying that your gear is connected (like some
quick-access actions, perhaps a couple of moves, a toggle for casual
mode, that sort of thing) would be very nifty indeed.