Introduce IsKnown and AutoConnect concepts, and honour them
IsKnown is set to true after the first connection to a previously not
known device. It is currently impossible to "forget" a device again, but
that'll come shortly.

AutoConnect is set to the same value as the global auto reconnect
option's value when a device is connected to. When a device with this
option set to true is discovered, connection is performed as soon as it
is added to the model.