Adjust logic to not offer light commands if device says it doesn't have lights
This affects the MiTail Mini for the most part, as it previously would
just outright not show the light commands even if it said it supported
the lights. So, that should be better now.

It also only loads the default commands with lights now if the device
says it has lights connected for the MiTail.